tests/wpt/web-platform-tests/css/css-flexbox/flexbox-overflow-auto-002.html

<!DOCTYPE html>
<html>
<title>CSS Flexbox: Height with overflow: auto.</title>
<link href="support/flexbox.css" rel="stylesheet">
<link rel="help" href="https://www.w3.org/TR/css-flexbox-1/#flex-direction-property">
<link rel="help" href="https://www.w3.org/TR/css-overflow-4/#classic-scrollbars">
<meta name="assert" content="This test ensures that flexbox with 'flex-direction: row|row-reverse' and a flex item child with 'overflow: auto' has the proper height."/>
<style>
.flexbox {
    border: 5px solid green;
    position: relative;
    width: 50px;
}

.inline-flexbox {
    border: 5px solid green;
    position: relative;
    height: 50px;
}

.overflow {
    border: 1px solid red;
    overflow: auto;
    min-width: 0;
    min-height: 0;
}

.vertical {
    writing-mode: vertical-rl;
}
</style>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/resources/check-layout-th.js"></script>
<body onload="checkLayout('.flexbox, .inline-flexbox')">
<div id=log></div>
<div class="flexbox" data-expected-height="47">
  <div class="overflow"><div style="width: 100px; height: 20px"></div></div>
</div>

<div class="flexbox row-reverse" data-expected-height="47">
  <div class="overflow"><div style="width: 100px; height: 20px"></div></div>
</div>

<div class="flexbox vertical" data-expected-height="47">
  <div class="overflow"><div style="width: 100px; height: 20px"></div></div>
</div>

<div class="flexbox row-reverse vertical" data-expected-height="47">
  <div class="overflow"><div style="width: 100px; height: 20px"></div></div>
</div>

<div class="flexbox" data-expected-height="47">
  <div class="overflow align-self-baseline"><div style="width: 100px; height: 20px"></div></div>
</div>

<div class="flexbox row-reverse" data-expected-height="47">
  <div class="overflow align-self-baseline"><div style="width: 100px; height: 20px"></div></div>
</div>

<div class="flexbox vertical" data-expected-height="32">
  <div class="overflow align-self-baseline"><div style="width: 100px; height: 20px"></div></div>
</div>

<div class="flexbox row-reverse vertical" data-expected-height="32">
  <div class="overflow align-self-baseline"><div style="width: 100px; height: 20px"></div></div>
</div>

<div class="inline-flexbox column" data-expected-width="47">
  <div class="overflow"><div style="width: 20px; height: 100px"></div></div>
</div>

<div class="inline-flexbox column-reverse" data-expected-width="47">
  <div class="overflow"><div style="width: 20px; height: 100px"></div></div>
</div>

<div class="inline-flexbox column vertical" data-expected-width="47">
  <div class="overflow"><div style="width: 20px; height: 100px"></div></div>
</div>

<div class="inline-flexbox column-reverse vertical" data-expected-width="47">
  <div class="overflow"><div style="width: 20px; height: 100px"></div></div>
</div>

<div class="inline-flexbox column" data-expected-width="47">
  <div class="overflow align-self-baseline"><div style="width: 20px; height: 100px"></div></div>
</div>

<div class="inline-flexbox column-reverse" data-expected-width="47">
  <div class="overflow align-self-baseline"><div style="width: 20px; height: 100px"></div></div>
</div>

<div class="inline-flexbox column vertical" data-expected-width="32">
  <div class="overflow align-self-baseline"><div style="width: 20px; height: 100px"></div></div>
</div>

<div class="inline-flexbox column-reverse vertical" data-expected-width="32">
  <div class="overflow align-self-baseline"><div style="width: 20px; height: 100px"></div></div>
</div>
</body>
</html>
